funciona

+++ proyecto/src/main/java/com/javeriana/drpaws/proyecto/controlador/NotFoundException.java
@@ -1,17 +1,17 @@
 package com.javeriana.drpaws.proyecto.controlador;
 
 public class NotFoundException extends RuntimeException {
-    private int id;
+    private Long id;
 
-    public NotFoundException(int id) {
+    public NotFoundException(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
